HA-42 Frontend : 크루, 공고 양식 파일 분리
