+++ src/routes/CrewCreate.tsx
@@ -1,4 +1,4 @@
-import { CrewCreateForm } from "../components/InputForm";
+import { CrewCreateForm } from "../components/CrewForm";
 import { CrewCreateTitle } from "../components/Titles";
 
 const CrewCreate = () => {
